test nix funcs with pytest via nix_ffi.py
