+++ tests/unit/nix_ffi.py
@@ -1,0 +1,1 @@
+../../src/apps/cli/nix_ffi.py
